reduced the gSize constant to fix race detector errors
The race detector was occasionally complaining that checkGidOffset() was
reading from an offset that was concurrently written to by another
goroutine. These checks are benign and necessary, but they are
technically race conditions. (The library scans the memory to find
goroutine ids and will of course read wrong addresses.)

Reducing the gSize from 4096 to 256 fixed the race detector complaints.
If the goid is stored past offset 256 in the future, the unit tests will
break and the gSize constant will need to be upped.